% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pci.R
\name{pci}
\alias{pci}
\title{Calculates Proactive Conservation Index}
\usage{
pci(sp, var_out, var_in = NULL, weight_out = NULL, weight_in = NULL)
}
\arguments{
\item{sp}{character. Names of the taxa being evaluated.}

\item{var_out}{numeric. Threat variables. higher values must indicate increased threat.}

\item{var_in}{numeric. Interacting variables. Will modulate the effect of threat variables.}

\item{weight_out}{numeric. Weights for threat variables}

\item{weight_in}{numeric. Matrix of weights for the combination of interacting variables and threat variables.}
}
\value{
Data frame with PCI and rank.
}
\description{
\code{pci} Calculates the Proactive Conservation Index, a new tool to prioritize species for conservation, which incorporates information about future threats.
}
