% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reptile_df.R
\docType{data}
\name{reptile_df}
\alias{reptile_df}
\title{Threat data for 10851 global reptiles.}
\format{
A data frame with 359 rows and 3 variables:
\describe{
  \item{sp}{character. species name}
  \item{iucn_cat}{character. IUCN Red List threat category}
  \item{mass}{numeric. Maximum body mass, in grams}
  \item{insular}{binary. If insular endemic, value 1, if not, value 0}
  \item{range_area}{numeric. Area of distribution range, in km2}
  \item{human_footprint}{numeric. Average human footprint (ref) density across species distribution}
  \item{pop_dens}{numeric. Average human population density across species distribution}
  ...
}
}
\usage{
reptile_df
}
\description{
A dataset containing data on threat correlates for 10851 global reptiles
}
\keyword{datasets}
