% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_weights.R
\name{optim_weights}
\alias{optim_weights}
\title{Optimizes weighting for the calculation of Proactive Conservation Prioritization Index}
\usage{
optim_weights(
  sp,
  var_out,
  var_in = NULL,
  weight_out = NULL,
  weight_in = NULL,
  reference,
  type = "both",
  ...
)
}
\arguments{
\item{sp}{character. Names of the taxa being evaluated.}

\item{var_out}{numeric. Threat variables. higher values must indicate increased threat.}

\item{var_in}{numeric. Interacting variables. Will modulate the effect of threat variables.}

\item{weight_out}{numeric. Weights for threat variables}

\item{weight_in}{numeric. Matrix of weights for the combination of interacting variables and threat variables.}

\item{reference}{numeric. Threat rank towards which weights will be optimized.}

\item{type}{character. Optimize weights for threat variables ("out"), for interacting variables ("in") or for both ("both").}

\item{...}{additional arguments to be passed to function 'optim'.}
}
\value{
Vector ("out"), matrix ("in") or list ("both") with optimal weights.
}
\description{
\code{optim_weights} Optimizes weights for calculating Proactive Conservation Prioritization Index (ref),
}
\details{
The Pearson correlation between the calculated rank and 'reference' is displayed as the weights are optimized.
}
\examples{

# Load data
data(reptile_df)

# Calculate inverse range area
reptile_df$inv_range_area <- 1/(reptile_df$range_area)

# Remove 'DD' and 'NE' species
reptile_df_iucn <-
   reptile_df[!reptile_df$iucn_cat \%in\% c("DD", "NE"),]

# Change order of levels
reptile_df_iucn$iucn_cat <-
  factor(reptile_df_iucn$iucn_cat,
         levels = c("LC", "NT", "VU", "EN", "CR"))

# Convert IUCN categories to numeric
reptile_df_iucn$iucn_cat <- as.numeric(reptile_df_iucn$iucn_cat)

var_out_iucn <- reptile_df_iucn[,6:8]
var_in_iucn <- reptile_df_iucn[,3:4]

# Optimize weights
optim_weights_both <-
  optim_weights(sp = reptile_df_iucn$sp,
                var_out = var_out_iucn,
                var_in = var_in_iucn,
                reference = reptile_df_iucn$iucn_cat,
                type = "both",
                control = list(maxit = 5))

}
